Create processors from clojure, wherever, whenever you like.
  (processor "/queue/biscuit" #(publish "/queue/ham" (.toUpperCase %)))

We depend on the closer to close our connection at undeployment, but
there's a race condition where a destination might shut down before
the connection is closed. We can move Closer to polyglot and make the
destination services depend on him to fix *or* we can create
destinations in clojure instead.