Got infinispan backing core.memoize finally.
This was tricky, and probably needs a lot more testing. It's tricky
for two main reasons: 1) memoize caches delay instances, which aren't
serializable, so we maintain a kind of L1 cache storing puts (which we
want to happen without incurring the potential cost of forcing the
delay) and we transfer the data from the L1 to infinispan on the
lookup, and 2) because we encode the keys and values of our cache, the
automatic implementations defined for us by defcache don't
work. Ideally, Clojure would let me tweak the impls of a defined type,
but there doesn't seem to be any way to do it.