Give each ClojureRuntime its own thread pool.
This prevents the memory leak from the ThreadLocals that Clojure uses. We
need our own thread pool because otherwise we end up on MSC threads that
get recycled between deploys, which end up holding a lot of memory.

There may be performance implications of this change - I'd love to talk to
@bbrowning and @jcrossley about it.