Incorporating Rich/Stu's feedback, no longer mutating the cache with core fns.
The InfinispanCache should now respond to the core functions in the
same way a java.util.HashMap would.

The confusion over Fogus requiring a clojure map-like thing came from
me trying to replicate his unit tests, including those for
assoc/dissoc, but that's really more a test of the defcache macro, not
CacheProtocol. So it's not an appropriate test for a mutable cache
that can still act as a memoization backing store.

Also added a delete-all function to Mutable that behaves like
Map#clear but returns the instance instead of void.