An interesting approach to in-container testing
We basically deploy an app to create and run some tests and if they
all pass, we start up a dumb web handler to accept a request. The
chief problem with this approach is that all useful info regarding the
in-container tests is in the server log. But I'm willing to trade that
inconvenience for transaction integs for now.