Remove the hacky handler that cleared the ring-session cookie.
We had this to prevent duplicate cookies (one for JSESSIONID, one for
ring-session, both with the same value), but it turns out ring needs
the ring-session value to properly clear sessions, even when we are
using ServletSession. My last change here had the handler stuffing
a ring-session cookie back in place, which is brittle and foolish.
So, we're ditching this handler altogether.