Removed the pojo subsystem to fix [IMMUTANT-144]
The pojo subsystem apparently tries to deploy beans it finds in any
jboss-beans.xml file it finds on the classpath. The netty jar has one,
of course, but it references a bean that's already installed in the
AS, because it depends on netty, too.

A side benefit of removing the pojo subsystem is the disappearance of
this annoying log message: "Found legacy bean/pojo namespace:
urn:jboss:bean-deployer:2.0 - might be missing some xml
features (potential exceptions)."

It remains to be seen whether Immutant users will rely on this
now-missing feature. I kinda doubt it. :)