Making web/start support a subset of ring-server's opts [IMMUTANT-241]
Breaking change is that :reload option is renamed :auto-reload?, but
that was pretty broken anyway since it hard-coded the :reload-paths
option to the project's "src" dir. We now add all directories on the
CP, by default.

Also, we detect "development mode" either by LEIN_NO_DEV being unset
(you might set it prior to starting Immutant) or the presence of :dev
as an active profile in the deployed project. Both :auto-reload? and
:stacktraces? default to true in development mode, false otherwise.