Fixes IMMUTANT-258 finally!
This means invoking immutant.cache/cache twice for the same-named
cache blows away whatever data is in there and reseeds it.

But I think that's what most folks would expect. It also means that
the options in the last call are what should remain in effect for
subsequent use. This wasn't the case up to now. Only changing the
replication mode option actually caused the cache to restart.