Added functions for setting session timeout and cookie options [IMMUTANT-267]
This adds set-session-timeout! for overriding the internal session
timeout, and set-session-cookie-attributes! to override the settings
for the session cookie.

We need to do it this way instead of by interpreting the ring session
arguments, since changing the session settings is per-deployment, not
per-endpoint, and expecting the per-endpoint session options to work
consistently would lead to hard to debug issues for the app developer.

Since you can now change the session cookie name, I'd like to look at
overriding the ring-session cookie completely.