Activate [:dev :base :user] profiles by default [IMMUTANT-270]
We're doing this to make development easier - it brings in checkout
deps, and allows you do to things like:

    :user {:immutant {:nrepl-port 0}}

in profiles.clj to get an nrepl automatically for every app you
deploy.

This also includes a notice in the docs about *not* using the default
profile set in production.

Let us never speak of this again.