Allow jobs to be unscheduled immediately after scheduling [IMMUTANT-243]
This issue was caused by the job being created asynchronously, and
stop checking to see if the job was started before trying to stop. It
will now check to see if the job is pending (thanks to
https://github.com/projectodd/jboss-polyglot/commit/526dd8517d9c856e702cb5c454278d6120abed31),
and wait for the job to reach the started state before actually
stopping it. Depending on the schedule for the job, it may actually
fire before being stopped.