Significant changes to support all versions of java.jdbc [IMMUTANT-277]
The main thing we did is remove the "transaction strategy injection"
hack, preferring instead to return a reified Connection that eats the
illegal XA calls (setAutoCommit, commit, and rollback) and satisfies
enough of the the Connection interface to appease the three major
versions of java.jdbc.

One big benefit of this is the removal of the java.jdbc dep from
immutant.xa.

I also rearranged the tests a good bit to better reflect which version
they're testing.