Correct my misunderstanding of useSynchronization to fix [IMMUTANT-335]
The only way Infinispan will register itself as an honest-to-goodness
XAResource participant is if useSynchronization is false. It's true by
default, useful for say, a Hibernate 2L cache, but not good when other
XA participants depend on the cache's state.

So now we always set useSynchronization to false when inside the
container.

Integration test provided as well.