Upgrade integs to use java.jdbc 0.3.2 [IMMUTANT-389]
As of 0.3.0, java.jdbc added support for isolation levels, exposing some
missing methods in our Connection wrapper. We now delegate those,
eliminating the AbstractMethodError. It remains to be seen how setting
an isolation level in an XA tx will work with various drivers. I don't
expect it to be a common use case.