Fixing namespaces
Firstly, we're no longer cheating by relying on the corresponding module
as a dependency, as it won't be available in standard repos.

Nextly, we're using the actual project.clj files generated by the
generate_clojars_project.clj script, with provided deps for the modules
to ensure proper build order and a hook that copies over the module jar
from a relative path.