Running generate-build-info.rb successfully now
This actually results in a fully-operational assembly!

Unfortunately, the versions hook is applied to all projects, when it
should really just be modules/core's. This is part of a larger issue
where sub-projects aren't run in their own isolated subprocess. This
also results in problems when a cwd of the child project's root is
assumed.