Replaced clj-http with http.async.client for testing [IMMUTANT-456]
My goal was to replace gniazdo with it, too, since http.async.client
should be able to do normal HTTP as well as WebSockets, but its
websocket close function is broken. Its websocket create function was
also limited in that I couldn't pass cookies to it in order to confirm
access to an Endpoint's Servlet's HttpSession, so I fixed that and
released my own fork to clojars.