Bundle tools.nrepl in the bootstrap module [IMMUTANT-459]
We activation lein plugin middleware when reading a project in the
bootstrap runtime, and cider.nrepl loads a tools.nrepl namespace upon
activation. Normally, tools.nrepl will be available when running
normally under lein, so cider.nrepl does not have a runtime dependency
on it. However, in the bootstrap runtime, tools.nrepl won't be available
unless we explicitly bring it in.