Further performance improvements for HttpHandlers and Servlets.
This restores the HttpHandler perf loss in the last commit, and
increases their perf above what they were originally. This also improves
on Servlet perf over the last commit.

This utilizes a HashMap as the basis for the ring request LazyMap to
save the cost of building and assoc'ing onto a
PersistentHashMap. Modifications to that hash (via assoc, dissoc) create
a PersistentHashMap instead.

For numbers, see: https://gist.github.com/9b9536773ce5ff54d642

* web/src/immutant/web/internal/ring.clj: