Redo integs so that there's just one project
The integs project is now always a part of the root build, but unless
the :integs profile is active, it's a no-op. The :integs profile is now
a composite of the integs for each library, which are the same tests in
the source for the library but run inside a WildFly container.

Integs for a single lib, say messaging, can be run inside the integs
project like so: lein with-profile +messaging all