Moving pieces in position for in-container tx testing
Unfortunately, we get strange errors and warnings involving "HQ222176: A
session that was already doing XA work" and invalid
XID's (XAException.XAER_NOTA).

Also, we need deploy-tools to deal with the new tx module. It'd be swell
if that wasn't necessary, but I guess the frequency of adding modules
will be low enough.