Allow records to be passed through a destination with :none encoding [IMMUTANT-468]
When :none is used, HornetQ uses java serialization, and deserialization
uses the TCCL to load classes. This ensures that the TCCL is set to
RT.baseLoader for listener threads, and for receives at the end of a
pipeline.

Users using receive directly with :none may have to set the TCCL
manually to avoid an error when decoding records.