Support for session cookie attributes out-of-container [IMMUTANT-510]
In-container handlers become servlets, and cookie attrs are configured
via the servlet context, which must be done during its init. But there
is no way to access that context from within a ring middleware function.

Two alternatives come to mind: 1) provide an imm1-style
set-session-cookie-attributes! that can only be invoked in main (but
finding the context may still be tricky) or b) use the cookie-config
element in web.xml.