Run faster with lein 2.5.1+ now
Mainly by updating to lein-modules 0.3.10, which uses the faster
read-raw function introduced in 2.5.0.

But also by setting :pedantic? to false in all the modules as its
default value changed from nil to 'ranges, which breaks memoization of
project dependencies, which slows lein-modules down a lot.

To get the pedantic check, activate the new +pedantic profile, which may
want to do on CI.