Ensure :port 0 returns the actual port in result [IMMUTANT-519]
This is somewhat tricky since we use the :port in the server name to
obtain the wunderboss Web instance. So we need the actual port in that
name. Hence, we ask the OS for an available port by briefly opening
a ServerSocket at port 0. This is gross but simple, and hopefully
effective for something that should be a rare, developer-only use
case.