Remove WS middleware order restriction out-of-container
With this change the wrap-websocket middleware no longer has to appear
at the end of the middleware chain when running out-of-container. It can
go anywhere that makes sense, as its result is no longer an HttpHandler,
but a simple function.

The must-come-last restriction is still in effect in-container, however,
as the result is then a Servlet rather than a function.