Fix context/path-info for upgrades [IMMUTANT-521]
This uses reflection, which isn't ideal. You only pay for it if you need
it, since request entries are delays, but the larger issue is the
reflection in the servlet impl ties us to specific data members in
Undertow classes. We can probably come up with an extension of
HandshakeRequest in wboss to mitigate this, though.