Ported over clustering tests
In the integration-tests dir, run:
  lein with-profile +cluster all

The old tests were in individual files, now they're all in
cluster_test.clj. All the server-side resources are invoked via
src/integs/cluster.clj, so it's the integration-tests app itself
deployed by fntest. We do this mainly because we need an app that
depends on the current version of immutant (for war creation).
Therefore, we :eval-in :leiningen to ensure all resources necessary to
read this project are in the classpath.

There seems to be an issue with remote messaging so the tests that rely
on that are currently commented.