Fix jersey dependencies for integs
Now that the dev war includes all of the deps (instead of just pointing
to them on the classpath), we have to exclude jersey-server to prevent
"JBAS011232: Only one JAX-RS Application Class allowed." exceptions on
deployment. The exclusions get us closer to being able to run the sse
tests in-container, but RESTEasy still isn't quite happy, throwing
"java.lang.IllegalArgumentException: Unable to find a public constructor
for provider class org.glassfish.jersey.media.sse.OutboundEventWriter",
so the sse tests are still disabled in-container.