No longer recursively munging option map keys [IMMUTANT-604]
I'm not sure why we ever thought this was a good idea, probably just
never thought any option values would be maps. But now they are, and we
really shouldn't be messing with them.

One bit of fallout from this is message metadata, where keyword keys
used to be turned into Strings (as they must be valid Java identifiers
to be used in selectors) automatically. Now they're not. An error will
occur if you attempt to pass keyword keys as metadata. I think this
probably better satisfies the "principle of least surprise"