Don't always assume websockets are available [IMMUTANT-600]
We blindly assumed the javax.websocket.server.ServerContainer was always
available, but it will only be available in EAP6 if websockets have been
configured in standalone.xml *and* in a jboss-web.xml in the WAR.

We can't know if a user's code will use websockets, so we have to try to
set up the endpoint for every servlet deployed. That means that users
that had no interest in websockets (and hadn't configured them) would
still be unable to run any ring handlers.

This changes the servlet init code to confirm the ServerContainer is
available, logging if it isn't found instead of throwing an NPE.