Use at-exit's new shutdown hook [IMMUTANT-605]
Confirmed that the method calls on the GracefulShutdownHandler are
idempotent, so no problem if the server is stopped prior to shutdown. If
a user wants to drain requests after stopping but prior to shutdown,
they must invoke the methods manually (on the object returned from
graceful-shutdown, which they ostensibly passed to web/run)

I considered adding an option to web/run for a "drain timeout" and
always wrapping the handler in a GracefulShutdownHandler. But since this
is only relevant for Undertow itself, i.e. out-of-container, I thought
it best to keep the timeout as a parameter to a separate function.