Return :port and :host from undertow/options [IMMUTANT-612]
Because it's difficult to tell exactly which listeners are registered
with the Undertow/Builder associated with the :configuration key, not to
mention that there may be multiple ports configured, the best we can do
I think is to return the most recently passed port. It's really either
that or return no port at all, which for the common case -- one port --
isn't really helpful -- especially when the port is 0.

Regardless, returning the default port, 8080, after Undertow-specific
options are passed is just dead wrong, so we fixed that, at least.