Allow proper aborting of websocket connection [IMMUTANT-632]
This brings in a new wunderboss that allows returning a non-async
body (and your own status code) to abort the websocket connection.

There are no changes in Immutant itself to support this out of
container, but for in container, we have to throw an exception from
within Immutant at the appropriate time to abort the connection.